+++ Food.Api/Models/AccountViewModels.cs
@@ -42,7 +42,7 @@ namespace Food.Api.Models
 
         public string LoginProvider { get; set; }
 
-        public int Balance { get; internal set; }
+        public decimal Balance { get; internal set; }
 
         public ICollection<IdentityUserClaim> Claims { get; internal set; }
         public DateTime? LockoutEndDate { get; internal set; }
